+++ CHANGELOG.md
@@ -21,6 +21,7 @@ to [Semantic Versioning]. The full commit history is available in the [commit lo
 - Fix inference on GPU for PyTorch implementation of {class}`scvi.external.MRVI`, {pr}`3586`.
 - Fix in non-multi-GPU training to have history in memory, and not on disk by default {pr}`3543`.
 - Fix missing model history for multi-GPU training, and add an option to log on step {pr}`3516`.
+- Fix external indices validation in {class}`scvi.dataloaders.SemiSupervisedDataSplitter` {pr}`3601`.
 
 #### Changed